step 2: complete
Created API helper.  This includes the functions that the frontend will
use to call the backend.

Since Vite is used for the frontend, Node's process module will be
undefined when running the frontend.  As a result, to have the BASE_URL
variable be correctly assigned for all circumstances, there has to be
checks for whether process or import.meta is available.

Also updated Axios import path.