step 6: feat: create SignUp
React Component Hierarchy is updated according to new info from
instructions.  The sign up form does not need to set the "user" state
in App, it only needs to have App do the setting and retrieving of
user.  App will get user info when the token changes.

In the App component, for rendering the NavBar, the isLoggedIn property
is changed to false to help with manual testing.  This value will
eventually be changed to be dynamic.