step 6: feat: update NavBar
Updated NavBar to display the current user's name and an option to log
out.  There was no need to have isLoggedIn if username can determine
that as well.

Instead of receiving a prop function that deletes the authentication
token, NavBar now receives a prop function for logging out.  The logic
for logging out will be given to the App component.

Also added in tests.