step 7: feat: create AllRoutesWrapper
Created a route wrapper that wraps all the other routes and that is
essentially a middleware that is run before rendering any page
component.  This AllRoutesWrapper component can contain any type of
code, but is currently used to protect routes from users who are not
logged in.

An alternative to this could be to conditionally render routes.