fix: add error handling for JoblyApi method calls
Calls to the API from pages, like CompaniesPage, CompanyPage, and
JobsPage, had no error handling.  This meant that issues with
retrieving company or job data would not be properly conveyed to users.

All pages have been updated to do basic error handling, which is to
display the error message on to the page.  Other uses of JoblyApi
methods, like in App, are not handled, because there is nothing that
can be done there.

Testing is skipped for now due to time.