step 9: feat: allow applying to jobs
In App, created a new function, applyToJob, to handle the functionality
of applying to a job.  This replaces the original intention of using
the state function, setApplications.  The hierarchy tree has been
updated accordingly.

JobCard has been updated to use the new applyToJob function.  As a
result, the testing for JobCard now uses mocking for useContext to
retain unit testing as much as possible.

JobsPage has been updated to now correctly display which jobs were
applied to.  New tests are added for displaying the correct "apply"
button and for testing the apply functionality.  React context
providers are used in tests to retain integration testing as much as
possible.

Also removed the unused mocking of react-router-dom in the JobsPage
test file.

CompanyPage has been updated to not have applications be passed in as a
prop, but be passed in from a React context.  This should tidy up any
rendering calls for CompanyPage.  CompanyPage tests are updated
similarly to JobsPage.

On another note, for other page components, like HomePage, data and
functions will continue to be passed in through props.  This is due to
time constraints.  In the future, they might utilize useContext.