chore: move code out of frontend folder
Originally, the backend and frontend code was in a single repository,
so it was in a separate folder.  Now since the frontend code is in its
own repository, it doesn't need the folder anymore.

The React component hierarchy is moved to the root.  api.js is moved to
src.

Imports are updated.  Extensions are added to all imported files, so
that they are consistent and leaves no possibility of compatibility
issues (not recognizing a .jsx file).

There were 2 .gitignore files: from GitHub and from Vite.  The Vite's
.gitignore is kept and the one from GitHub has been deleted.