refactor: remove logging in for App tests
In App.test.jsx, some tests were manually logging in a user in order to
perform the main tests, like for viewing job openings.  This was
unnecessary as logging in is not the focus of the tests, and logging in
already has a test.

The code for manually logging in has been replaced with setting the
authToken in local storage.  This helps to reduce unneeded, duplicate
code.