fix: correct base URL in api.js
The deployed frontend was calling the backend at the URL
localhost:3001.  This indicates that the environment variable
REACT_APP_BASE_URL was not being read.

The code for getting the BASE_URL has been simplified to try to resolve
this.  Since Vite does not use process.env during production, the
checks for that has been removed.