feat: fix up .body, #root, and .App CSS
Cleared out some old Vite settings and simplified .body, #root, and
.App CSS.

.body has minimum height set to 100% of view height, so that everything
can be vertically centered.  The rest of the properties are no longer
necessary, as they are not being utilized.

For #root, the max-width is too restrictive, and margin and padding are
no longer relevant.

For .App, "display: flex" and "min-height: 100vh" are needed to help
center page content that is shorter than the viewing screen height,
such as the home and profile pages.

.card CSS properties are removed, because that class is not being used.

NavBar is given a flex CSS property, so that any HTML elements under it
can expand to fill up the rest of the viewing screen.  This helps with
centering content.