feat: add reactstrap for HomePage and form pages
Added Reactstrap for HomePage, Login, SignUp, and Profile.  Most of the
CSS styling had to be placed in App.css, as multiple elements shared
the same CSS.

Removed error and feedback message divs and replaced them with
Reactstrap's Alert.

In Profile, feedbackMessage's data type is changed from String to
Object, in order to hold the feedback type.  The type is used to
display the correct color of the feedback, as it can be for a success
or fail.